General Organisation
Improved commenting and also added a file of helper functions for the command interface to clean up the class a bit. Unfortunately despite their impact on the readability the parsers cannot be moved to this helper functions file which was the original intent. However, the class has still been improved and simplified.